requirements.txt
# Keep in sync with setup.py's install_requires option
# https://packaging.python.org/discussions/install-requires-vs-requirements
apispec==0.39.0
configfile==1.2.2
Flask-Cors==3.0.6
flask-marshmallow==0.9.0
Flask-Migrate==2.2.1
flask-restinpeace==1.2.0
Flask-SQLAlchemy==2.3.2
marshmallow-sqlalchemy==0.14.1
pyxdg==0.26
